[#48] Refactor: DB 설정 정보 core 모듈로 이동
